### 2018-11-12 #1078 @whedon
[PRE REVIEW]: IDTxl: The Information Dynamics Toolkit xl: a Python package for the efficient analysis of multivariate information dynamics in networks
**Submitting author:** @pwollstadt (<a href="http://orcid.org/0000-0002-7105-5207">Patricia Wollstadt</a>)
**Repository:** <a href="https://github.com/pwollstadt/IDTxl" target ="_blank">https://github.com/pwollstadt/IDTxl</a>
**Version:** v.0.4.0
