### 2018-11-14 #1078 @jlizier
Hi -- some reviewer suggestions from @pwollstadt and I are as follows:
1. Daniele Marinazzo @danielemarinazzo who is familiar with the inference algorithms and underlying information theory measures here.
2. Ryan James @Autoplectic who is familiar with the underlying information theory measures and has published with JOSS in the past.
3. Luca Faes (can't find a github account though), whose expertise here is similar to Marinazzo
