### 2018-11-14 #1078 @danielemarinazzo
good, hit me. If you preprint I can do an open review too.
I won't be able to review the code though, you might want someone expert in
that.

Luca Faes has no github account indeed afaik.

cheers

On Wed, Nov 14, 2018 at 1:01 PM Joseph Lizier <notifications@github.com>
wrote:

> Hi -- some reviewer suggestions from @pwollstadt
> <https://github.com/pwollstadt> and I are as follows:
>
>    1. Daniele Marinazzo @danielemarinazzo
>    <https://github.com/danielemarinazzo> who is familiar with the
>    inference algorithms and underlying information theory measures here.
>    2. Ryan James @Autoplectic <https://github.com/Autoplectic> who is
>    familiar with the underlying information theory measures and has published
>    with JOSS in the past.
>    3. Luca Faes (can't find a github account though), whose expertise
>    here is similar to Marinazzo
>
