### 2018-11-14 b931649
Fix multiple citations and add missing author
