+++ paper/paper.bib
@@ -1,6 +1,6 @@
 @book{Bossomaier2016,
 address = {Cham, Switzerland},
-author = {Bossomaier, Terry and Barnett, Lionel and Lizier, Joseph T.},
+author = {Bossomaier, Terry and Barnett, Lionel and Harr{\'e}, Michael and Lizier, Joseph T.},
 publisher = {Springer},
 title = {An Introduction to Transfer Entropy},
 note = {\url{https://doi.org/10.1007/978-3-319-43222-9}},
