### 2018-11-14 #1078 @pwollstadt
Hi all!

@cMadan, we have reviewed the pdf and have found a problem with multiple citations. I fixed that and uploaded the corrected version (https://github.com/pwollstadt/IDTxl/commit/b9316494054e82cff8342e54bc5189a05f677fb4). There was another error in a reference, I also got rid of. Is there any action required on my part to recompile the pdf?

Thanks @jlizier for adding our suggested reviewers and thanks already to @danielemarinazzo for the quick reply!
