### 2018-11-14 #1078 @jlizier
The 2nd generated PDF proof is fine.

(@cMadan I hope I'm not out of line here by adding:) @danielemarinazzo I don't think JOSS is too worried about code level review, more a check of required features such as documentation, tests, etc -- see https://joss.readthedocs.io/en/latest/reviewer_guidelines.html and see the checklist for the reviewer in issue #1075 for example (a similar checklist would be added here once reviewers are formally assigned). Thanks for the quick reply!
