### 2018-11-15 #1078 @cMadan
@jlizier, no problem about revising the PDF. I wouldn't quite agree that we're not worried about code-level review. It's definitely not line-by-line, but this is a software journal and part of evaluating the functionality is agreeing that the code as a whole does what it says it's supposed to do.

@danielemarinazzo, it would be great to have you as a reviewer for this. Evaluate what you think you can, and let me know what aspects you feel unsure to judge--and then we'll figure things from there.

@Autoplectic, great, thanks for volunteering!
