### 2018-11-15 5acbb2f
Add missing dois, fix inline citations
