### 2018-11-15 428336a
Remove url command from dois
