+++ paper/paper.bib
@@ -3,13 +3,13 @@ address = {Cham, Switzerland},
 author = {Bossomaier, Terry and Barnett, Lionel and Harr{\'e}, Michael and Lizier, Joseph T.},
 publisher = {Springer},
 title = {An Introduction to Transfer Entropy},
-doi = {\url{https://doi.org/10.1007/978-3-319-43222-9}},
+doi = {https://doi.org/10.1007/978-3-319-43222-9},
 year = {2016}
 }
 @article{Faes2011,
 author = {Faes, Luca and Nollo, Giandomenico and Porta, Alberto},
 doi = {10.1103/PhysRevE.83.051112},
-doi = {\url{https://doi.org/10.1103/PhysRevE.83.051112}},
+doi = {https://doi.org/10.1103/PhysRevE.83.051112},
 journal = {Physical Review E - Statistical, Nonlinear, and Soft Matter Physics},
 pages = {1--15},
 title = {{Information-based detection of nonlinear Granger causality in multivariate processes via a nonuniform embedding technique}},
@@ -20,7 +20,6 @@ year = {2011}
 author = {Kraskov, Alexander and St{\"{o}}gbauer, Harald and Grassberger, Peter},
 doi = {10.1103/PhysRevE.69.066138},
 journal = {Physical Review E - Statistical, Nonlinear, and Soft Matter Physics},
-doi = {\url{https://doi.org/10.1103/PhysRevE.69.066138}},
 pages = {066138},
 title = {{Estimating mutual information.}},
 volume = {69},
@@ -30,7 +29,7 @@ year = {2004}
 author = {Lizier, Joseph T. and Rubinov, Mikail},
 journal = {Max Planck Institute: Preprint},
 title = {{Multivariate construction of effective computational networks from observational data}},
-doi = {\url{http://www.mis.mpg.de/preprints/2012/preprint2012_25.pdf}},
+note = {http://www.mis.mpg.de/preprints/2012/preprint2012_25.pdf},
 volume = {Preprint no. 25},
 year = {2012}
 }
@@ -43,7 +42,7 @@ year={2010}
 }
 @article{Montalto2014,
 author = {Montalto, Alessandro and Faes, Luca and Marinazzo, Daniele},
-doi = {\url{https://doi.org/10.1371/journal.pone.0109462}},
+doi = {https://doi.org/10.1371/journal.pone.0109462},
 journal = {PLoS ONE},
 number = {10},
 pages = {e109462},
@@ -53,7 +52,7 @@ year = {2014}
 }
 @article{Schreiber2000,
 author = {Schreiber, Thomas},
-doi = {\url{https://doi.org/10.1103/PhysRevLett.85.461}},
+doi = {https://doi.org/10.1103/PhysRevLett.85.461},
 journal = {Physical Review Letters},
 number = {2},
 pages = {461--464},
@@ -63,7 +62,7 @@ year = {2000}
 }
 @article{Vicente2011,
 author = {Vicente, Raul and Wibral, Michael and Lindner, Michael and Pipa, Gordon},
-doi = {\url{https://doi.org/10.1007/s10827-010-0262-3}},
+doi = {https://doi.org/10.1007/s10827-010-0262-3},
 journal = {Journal of Computational Neuroscience},
 number = {1},
 pages = {45--67},
@@ -73,7 +72,7 @@ year = {2011}
 }
 @article{Wibral2013,
 author = {Wibral, Michael and Pampu, Nicolae and Priesemann, Viola and Siebenh{\"{u}}hner, Felix and Seiwert, Hannes and Lindner, Michael and Lizier, Joseph T. and Vicente, Raul},
[…]
