### 2018-11-22 #1081 @danielemarinazzo
I have completed the checklist. Unless I have missed, and I apologize in advance, the license file seems to be missing.

Suggestions:

1. it would be convenient to add some running time indicators for different possible scenarios (number of time series, length of time series, etc).
2. you call "multivariate" and "bivariate" the cases with more or two time series respectively. In other contexts "multivariate" refers to the case in which the driver, or the target, or both, are comprised of many time series. On the other hand the case in which you want to distinguish among direct and mediated influences is called "conditioned" as opposed to "pairwise".  I understand that this is solely a matter of nomenclature, still a note of disambiguation could be useful.
3. In the examples you refer to "MuTE network" or "generate.mute.example".  If MuTE is  a general acronym for multivariate transfer entropy, no problem. If on the other hand it refers to the examples implemented in our toolbox (https://github.com/montaltoalessandro/MuTE), I would just like to stress the benchmark model with 5 time series is not ours but is taken by other toolboxes, such as the MVGC by Barnett and Seth, and before that the work by Baccala and Sameshima.

Congratultions and thanks for this nice upgrade.

d.
