### 2018-11-22 #1081 @cMadan
It looks like the license is in `COPYING.txt`. Is there a particular reason that this is the filename, rather than `LICENSE.txt`?
