### 2018-11-22 #1081 @jlizier
From memory that was to follow the [GNU recommendation](http://www.gnu.org/licenses/gpl-howto.html) - is the use of LICENSE mandatory for JOSS or other options ok so long as it's following some convention such as GNU?

@danielemarinazzo thanks for the comments - all are clear, we'll get back to you on those
