### 2018-11-23 #1081 @cMadan
JOSS doesn't have any rules about the license filename, but my impression is that `LICENSE` is more common of a filename than `COPYING`, even for GNU GPL licensed projects. The main idea is that the license should be easily found. I recommend you change the filename here, but I am not going to require it.
