### 2018-11-24 96a179f
Add class to handle adjacency matrices
Add class to handle adjacency matrices. Zero-lag analysis may result
in networks that contain edges with a delay of zero. The previous
handling of adjacency matrices as numpy arrays where non-zero entries
represent weighted edges didn't allow for edges with a delay of
zero. The AdjacencyMatrix class contains a boolean matrix for edges
and a weight matrix of varying type, allowing for weights of value
zero.

Adapt unit tests and functions expecting adjacency matrices as
input accordingly. Update test data.

Fixes #20. Fix minor bugs in network plotting.
