+++ test/test_results.py
@@ -6,6 +6,7 @@ from tempfile import TemporaryFile
 import itertools as it
 import copy as cp
 import numpy as np
+from idtxl.results import AdjacencyMatrix
 from idtxl.multivariate_te import MultivariateTE
 from idtxl.bivariate_te import BivariateTE
 from idtxl.multivariate_mi import MultivariateMI
@@ -128,13 +129,13 @@ def test_results_network_inference():
         assert res.data_properties.normalised == normalisation, (
             'Incorrect value for data normalisation.')
         adj_matrix = res.get_adjacency_matrix('binary', fdr=False)
-        assert adj_matrix.shape[0] == n_nodes, (
+        assert adj_matrix._edge_matrix.shape[0] == n_nodes, (
             'Incorrect number of rows in adjacency matrix.')
-        assert adj_matrix.shape[1] == n_nodes, (
+        assert adj_matrix._edge_matrix.shape[1] == n_nodes, (
             'Incorrect number of columns in adjacency matrix.')
-        assert adj_matrix.shape[0] == n_nodes, (
+        assert adj_matrix._edge_matrix.shape[0] == n_nodes, (
             'Incorrect number of rows in adjacency matrix.')
-        assert adj_matrix.shape[1] == n_nodes, (
+        assert adj_matrix._edge_matrix.shape[1] == n_nodes, (
             'Incorrect number of columns in adjacency matrix.')
 
 
@@ -249,10 +250,13 @@ def test_delay_reconstruction():
     res_network.combine_results(nw.analyse_single_target(
         settings=settings, data=data, target=3))
     adj_mat = res_network.get_adjacency_matrix('max_te_lag', fdr=False)
-    print(adj_mat)
-    assert adj_mat[0, 1] == delay_1, ('Estimate for delay 1 is not correct.')
-    assert adj_mat[0, 2] == delay_2, ('Estimate for delay 2 is not correct.')
-    assert adj_mat[0, 3] == delay_3, ('Estimate for delay 3 is not correct.')
+    adj_mat.print_matrix()
+    assert adj_mat._weight_matrix[0, 1] == delay_1, (
+        'Estimate for delay 1 is not correct.')
+    assert adj_mat._weight_matrix[0, 2] == delay_2, (
+        'Estimate for delay 2 is not correct.')
+    assert adj_mat._weight_matrix[0, 3] == delay_3, (
+        'Estimate for delay 3 is not correct.')
 
     for target in range(1, 4):
         est_mi = res_network._single_target[target].omnibus_te
@@ -331,50 +335,103 @@ def test_results_network_comparison():
     t = [1, 2]
     test = ['Within', 'Between']
     for (i, res) in enumerate([res_within, res_between]):
+        # Get adjacency matrices.
+        adj_matrix_union = res.get_adjacency_matrix('union')
+        adj_matrix_comp = res.get_adjacency_matrix('comparison')
+        adj_matrix_diff = res.get_adjacency_matrix('diff_abs')
+        adj_matrix_pval = res.get_adjacency_matrix('pvalue')
+
+        # Test all adjacency matrices for non-zero entries at compared edges
+        # and zero/False entries otherwise.
+
         # Union network
         # TODO do we need the max_lag entry?
-        assert (res.get_adjacency_matrix('union')[s, t] == 1).all(), (
+        assert adj_matrix_union._edge_matrix[s, t].all(), (
             '{0}-test did not return correct union network links.'.format(
                 test[i]))
-        no_diff = np.extract(np.invert(res.get_adjacency_matrix('comparison')),
-                             res.get_adjacency_matrix('union'))
-        assert (no_diff == 0).all(), (
[…]
