### 2018-11-25 1d61737
Rename COPYING to LICENSE
