### 2018-11-25 #1081 @pwollstadt
Hi @cMadan, I changed the file name to `LICENSE` in https://github.com/pwollstadt/IDTxl/commit/1d617372cca66449dd36139b3301c17391e193f7.

Thanks @danielemarinazzo for the quick review. I will look into the relevant references for the MuTE example, thanks for pointing that out.
