### 2018-11-26 #1081 @Autoplectic
Everything checklist-wise looks good to me.

A few minor suggestions:

You implement and document both [multiinformation](https://github.com/pwollstadt/IDTxl/wiki/Variable-names-used-in-IDTxl) as well as [multivariate mutual information](http://pwollstadt.github.io/IDTxl/html/idtxl_network_inference.html#multivariate-mutual-information-mmi). It would be best to utilize a single name, and to list alternative names (this quantity is often also called the total correlation, for example). Further, it would be good to explicitly state that which multivariate mutual informations it is *not* (co-information/integration, dual total correlation/binding information, [CAEKL mutual information](https://ieeexplore.ieee.org/document/7270410?reload=true&arnumber=7270410&filter%3DAND(p_IS_Number:7270357))).

To check my sanity, the SydneyPID estimator estimates the BROJA PID, not the recent PID of Lizier and Finn, yes?
