### 2018-12-02 55f672a
Remove redefinition of BROJA exception
Fixes #22.
