### 2018-12-02 65aa8e3
Update documentation
Update documentation. Add original reference for the example network
used in the MuTE paper and make clear that the Sydney PID estimator
estimates the BROJA measure. Add documentation for AdjacencyMatrix class.
