### 2018-12-12 21030cf
Fix typos in affiliations, fix doi links
