+++ paper/paper.bib
@@ -3,13 +3,12 @@ address = {Cham, Switzerland},
 author = {Bossomaier, Terry and Barnett, Lionel and Harr{\'e}, Michael and Lizier, Joseph T.},
 publisher = {Springer},
 title = {An Introduction to Transfer Entropy},
-doi = {https://doi.org/10.1007/978-3-319-43222-9},
+doi = {10.1007/978-3-319-43222-9},
 year = {2016}
 }
 @article{Faes2011,
 author = {Faes, Luca and Nollo, Giandomenico and Porta, Alberto},
 doi = {10.1103/PhysRevE.83.051112},
-doi = {https://doi.org/10.1103/PhysRevE.83.051112},
 journal = {Physical Review E - Statistical, Nonlinear, and Soft Matter Physics},
 pages = {1--15},
 title = {{Information-based detection of nonlinear Granger causality in multivariate processes via a nonuniform embedding technique}},
@@ -42,7 +41,7 @@ year={2010}
 }
 @article{Montalto2014,
 author = {Montalto, Alessandro and Faes, Luca and Marinazzo, Daniele},
-doi = {https://doi.org/10.1371/journal.pone.0109462},
+doi = {10.1371/journal.pone.0109462},
 journal = {PLoS ONE},
 number = {10},
 pages = {e109462},
@@ -52,7 +51,7 @@ year = {2014}
 }
 @article{Schreiber2000,
 author = {Schreiber, Thomas},
-doi = {https://doi.org/10.1103/PhysRevLett.85.461},
+doi = {10.1103/PhysRevLett.85.461},
 journal = {Physical Review Letters},
 number = {2},
 pages = {461--464},
@@ -62,7 +61,7 @@ year = {2000}
 }
 @article{Vicente2011,
 author = {Vicente, Raul and Wibral, Michael and Lindner, Michael and Pipa, Gordon},
-doi = {https://doi.org/10.1007/s10827-010-0262-3},
+doi = {10.1007/s10827-010-0262-3},
 journal = {Journal of Computational Neuroscience},
 number = {1},
 pages = {45--67},
@@ -72,7 +71,7 @@ year = {2011}
 }
 @article{Wibral2013,
 author = {Wibral, Michael and Pampu, Nicolae and Priesemann, Viola and Siebenh{\"{u}}hner, Felix and Seiwert, Hannes and Lindner, Michael and Lizier, Joseph T. and Vicente, Raul},
-doi = {https://doi.org/10.1371/journal.pone.0055809},
+doi = {10.1371/journal.pone.0055809},
 journal = {PloS ONE},
 number = {2},
 pages = {e55809},
@@ -86,7 +85,7 @@ doi = {10.1145/584091.584093},
 journal = {The Bell System Technical Journal},
 pages = {379--423},
 title = {A mathematical theory of communication},
-doi = {https://doi.org/10.1002/j.1538-7305.1948.tb00917.x},
+doi = {10.1002/j.1538-7305.1948.tb00917.x},
 volume = {27},
 year = {1948}
 }
@@ -102,7 +101,7 @@ year = {1948}
   author={Cover, Thomas M. and Thomas, Joy A.},
   year={2006},
   publisher={John Wiley \& Sons},
-  doi = {https://doi.org/10.1002/047174882X},
+  doi = {10.1002/047174882X},
   address={New York, NY}
 }
 @incollection{Wibral2014,
@@ -112,13 +111,13 @@ booktitle={Directed Information Measures in Neuroscience},
 editor={Wibral, Michael and Vicente, Raul and Lizier, Joseph T.},
 adress={Berlin, Heidelberg},
 publisher={Springer},
-doi = {https://doi.org/10.1007/978-3-642-54474-3_1},
[…]
