### 2018-12-12 #1081 @pwollstadt
Hi all, thanks to @danielemarinazzo and @Autoplectic for the review and the suggestions, we have addressed the remaining issues:

@danielemarinazzo, we have added the original reference for the example network used in the MuTE paper. Also, theoretical and some practical results for runtimes have been added as a [wiki page](https://github.com/pwollstadt/IDTxl/wiki/Runtimes-and-Benchmarking). I also added a sentence to the [theoretical introduction](https://github.com/pwollstadt/IDTxl/wiki/Theoretical-Introduction#idtxls-greedy-algorithm-for-mte-estimation) pointing our that what we call multivariate TE, is sometimes also termed "conditioned" TE. The major issue regarding the license file was already fixed earlier (see https://github.com/openjournals/joss-reviews/issues/1081#issuecomment-441458997).

@Autoplectic, we do not yet implement a measure for multiinformation, only multivariate mutual information. We apologise for the confusion. We have added a note to the [variable list](https://github.com/pwollstadt/IDTxl/wiki/Variable-names-used-in-IDTxl#idtxl-variable-and-function-names)  that it contains `multiinformation` for future reference and is not yet implemented. We hope, this will clear up misunderstandings about the implementation.

> To check my sanity, the SydneyPID estimator estimates the BROJA PID, not the recent PID of Lizier and Finn, yes?

Yes, The Sydney PID estimator estimates the BROJA PID. I added this information to the classes' docstring.

@editor, the commits also contain a fix for an issue raised independently of the JOSS review process ([#20](https://github.com/pwollstadt/IDTxl/issues/20)). I hope this is alright with everyone.

Thanks again to everyone for your suggestions so far!
