### 2018-12-12 #1081 @Autoplectic
@pwollstadt what exactly do you mean by multivariate mutual information? There are several quantities which have been called that, and so adding a definition or utilizing a more specific name would be very helpful.
