### 2018-12-19 #1081 @pwollstadt
@Autoplectic, we have added a description of our multivariate mutual information implementation to the [Wiki](https://github.com/pwollstadt/IDTxl/wiki/Theoretical-Introduction#further-network-inference-algorithms-implemented-in-idtxl). Thank you for your suggestion, we hope this clears up any misunderstanding about what is meant here.
