### 2018-12-23 370b94b
Fix style (whitespace, PEP8)
