+++ idtxl/network_inference.py
@@ -123,7 +123,7 @@ class NetworkInference(NetworkAnalysis):
                 #  we'll terminate the search for more candidates,
                 #  though those identified already remain valid
                 print('AlgorithmExhaustedError encountered in '
-                    'estimations: ' + aee.message)
+                      'estimations: ' + aee.message)
                 print('Halting current estimation set.')
                 # For now we don't need a stack trace:
                 # traceback.print_tb(aee.__traceback__)
@@ -136,14 +136,14 @@ class NetworkInference(NetworkAnalysis):
                 print('testing candidate: {0} '.format(
                     self._idx_to_lag([max_candidate])[0]), end='')
             try:
-                significant = stats.max_statistic(self, data, candidate_set,
-                                              te_max_candidate)[0]
+                significant = stats.max_statistic(
+                    self, data, candidate_set, te_max_candidate)[0]
             except ex.AlgorithmExhaustedError as aee:
-                # The algorithm cannot continue here, so
-                #  we'll terminate the check of significance for this candidate,
-                #  though those identified already remain valid
+                # The algorithm cannot continue here, so we'll terminate the
+                # check of significance for this candidate, though those
+                # identified already remain valid
                 print('AlgorithmExhaustedError encountered in '
-                    'estimations: ' + aee.message)
+                      'estimations: ' + aee.message)
                 print('Halting candidate max stats test')
                 # For now we don't need a stack trace:
                 # traceback.print_tb(aee.__traceback__)
@@ -521,7 +521,7 @@ class NetworkInferenceBivariate(NetworkInference):
                     #  we'll terminate the search for more candidates,
                     #  though those identified already remain valid
                     print('AlgorithmExhaustedError encountered in '
-                        'estimations: ' + aee.message)
+                          'estimations: ' + aee.message)
                     print('Halting current estimation set.')
                     # For now we don't need a stack trace:
                     # traceback.print_tb(aee.__traceback__)
@@ -539,10 +539,10 @@ class NetworkInferenceBivariate(NetworkInference):
                         te_max_candidate, conditional_realisations)[0]
                 except ex.AlgorithmExhaustedError as aee:
                     # The algorithm cannot continue here, so
-                    #  we'll terminate the significance check for this candidate,
-                    #  though those identified already remain valid
+                    # we'll terminate the significance check for this
+                    # candidate, though those identified already remain valid.
                     print('AlgorithmExhaustedError encountered in '
-                        'estimations: ' + aee.message)
+                          'estimations: ' + aee.message)
                     print('Halting candidate max stats test')
                     # For now we don't need a stack trace:
                     # traceback.print_tb(aee.__traceback__)
@@ -605,7 +605,8 @@ class NetworkInferenceBivariate(NetworkInference):
             [s[0] for s in self.selected_vars_sources])
         for source in significant_sources:
             # Find selected past variables for current source
-            print('selected vars sources {0}'.format(self.selected_vars_sources))
+            print('selected vars sources {0}'.format(
+                self.selected_vars_sources))
             source_vars = [s for s in self.selected_vars_sources if
                            s[0] == source]
             print('selected candidates current source: {0}'.format(
@@ -668,9 +669,9 @@ class NetworkInferenceBivariate(NetworkInference):
                     #  we'll terminate the pruning check,
                     #  assuming that we need not prune any more
                     print('AlgorithmExhaustedError encountered in '
-                        'estimations: ' + aee.message)
+                          'estimations: ' + aee.message)
                     print('Halting current pruning and allowing others to'
-                        ' remain.')
+                          ' remain.')
                     # For now we don't need a stack trace:
[…]
