### 2018-12-23 f79f4fd
Fix style (whitespace, PEP8)
