+++ idtxl/network_inference.py
@@ -137,15 +137,14 @@ class NetworkInference(NetworkAnalysis):
                     self._idx_to_lag([max_candidate])[0]), end='')
             try:
                 significant = stats.max_statistic(
-                    self, data, candidate_set, te_max_candidate,
-                    conditional=self._selected_vars_realisations)[0]
+                    self, data, candidate_set, te_max_candidate)[0]
             except ex.AlgorithmExhaustedError as aee:
-                # The algorithm cannot continue here, so  we'll terminate the
+                # The algorithm cannot continue here, so we'll terminate the
                 # check of significance for this candidate, though those
                 # identified already remain valid
-                print(
-                    'AlgorithmExhaustedError encountered in estimations: {}. '
-                    'Halting current estimation set.'.format(aee.message))
+                print('AlgorithmExhaustedError encountered in '
+                      'estimations: ' + aee.message)
+                print('Halting candidate max stats test')
                 # For now we don't need a stack trace:
                 # traceback.print_tb(aee.__traceback__)
                 break
@@ -522,10 +521,9 @@ class NetworkInferenceBivariate(NetworkInference):
                     # The algorithm cannot continue here, so
                     #  we'll terminate the search for more candidates,
                     #  though those identified already remain valid
-                    print(
-                        'AlgorithmExhaustedError encountered in estimations: '
-                        '{}. Halting current estimation set.'.format(
-                            aee.message))
+                    print('AlgorithmExhaustedError encountered in '
+                          'estimations: ' + aee.message)
+                    print('Halting current estimation set.')
                     # For now we don't need a stack trace:
                     # traceback.print_tb(aee.__traceback__)
                     break
@@ -541,13 +539,12 @@ class NetworkInferenceBivariate(NetworkInference):
                         self, data, candidate_set,
                         te_max_candidate, conditional_realisations)[0]
                 except ex.AlgorithmExhaustedError as aee:
-                    # The algorithm cannot continue here, so we'll terminate
-                    # the significance check for this candidate, though those
-                    # identified already remain valid
-                    print(
-                        'AlgorithmExhaustedError encountered in estimations: '
-                        '{}. Halting current estimation set.'.format(
-                            aee.message))
+                    # The algorithm cannot continue here, so
+                    # we'll terminate the significance check for this
+                    # candidate, though those identified already remain valid.
+                    print('AlgorithmExhaustedError encountered in '
+                          'estimations: ' + aee.message)
+                    print('Halting candidate max stats test')
                     # For now we don't need a stack trace:
                     # traceback.print_tb(aee.__traceback__)
                     break
@@ -712,10 +709,10 @@ class NetworkInferenceBivariate(NetworkInference):
                     # The algorithm cannot continue here, so
                     #  we'll terminate the pruning check,
                     #  assuming that we need not prune any more
-                    print(
-                        'AlgorithmExhaustedError encountered in estimations: '
-                        '{}. Halting current estimation set.'.format(
-                            aee.message))
+                    print('AlgorithmExhaustedError encountered in '
+                          'estimations: ' + aee.message)
+                    print('Halting current pruning and allowing others to'
+                          ' remain.')
                     # For now we don't need a stack trace:
                     # traceback.print_tb(aee.__traceback__)
                     break
@@ -754,9 +751,9 @@ class NetworkInferenceBivariate(NetworkInference):
             except ex.AlgorithmExhaustedError as aee:
                 # The algorithm cannot continue here, so
                 #  we'll set the results to zero
-                print(
-                    'AlgorithmExhaustedError encountered in estimations: {}. '
-                    'Halting current estimation set.'.format(aee.message))
+                print('AlgorithmExhaustedError encountered in '
[…]
