### 2019-01-14 #1081 @cMadan
@Autoplectic, do you have any additional questions, or are the definitions of the terms clear now? Thank you!
