### 2019-01-14 #1081 @Autoplectic
I have no further questions. With the multitude of multivariate information measures out there, it is important to have definitions as the same term is used to refer to different quantities rather often.
