### 2019-01-14 33eda1e
Fix bug in manual adding of conditioning variables
Fix bug in manual adding of conditioning variables. Convert
variables to lags bevor removing them from the candidate set.
Adjust unit tests accordingly.
