### 2019-01-17 #1081 @cMadan
@Autoplectic, of course, that completely makes sense.

@pwollstadt, it looks like we're all set to accept! Can you mint a new release that includes all of the changes from the review process, and provide a doi associated with an archived version on Zenodo or figshare?
