### 2019-01-31 #1081 @pwollstadt
@cMadan, I prepared a [final release](https://github.com/pwollstadt/IDTxl/releases/tag/v1.0) containing all requested changes and linked it to Zenodo. Here is the DOI: [10.5281/zenodo.2554339](https://doi.org/10.5281/zenodo.2554339). Let me know if there is anything more you need from us for the publication. Thanks again to @Autoplectic and @danielemarinazzo and everyone else for your input!
