+++ idtxl/bivariate_mi.py
@@ -140,6 +140,11 @@ class BivariateMI(NetworkInferenceMI, NetworkInferenceBivariate):
         assert(len(sources) == len(targets)), (
             'List of targets and list of sources have to have the same length')
 
+        # Check and set defaults for checkpointing. If requested, initialise
+        # checkpointing.
+        self.settings = self._set_checkpointing_defaults(
+            settings, data, sources, targets)
+
         # Perform MI estimation for each target individually
         results = ResultsNetworkInference(n_nodes=data.n_processes,
                                           n_realisations=data.n_realisations(),
@@ -244,6 +249,12 @@ class BivariateMI(NetworkInferenceMI, NetworkInferenceBivariate):
                   further settings (default=False)
                 - verbose : bool [optional] - toggle console output
                   (default=True)
+                - write_ckp : bool [optional] - enable checkpointing, writes
+                  analysis state to disk every time a variable is selected;
+                  resume crashed analysis using
+                  network_analysis.resume_checkpoint() (default=False)
+                - filename_ckp : str [optional] - checkpoint file name (without
+                  extension) (default='./idtxl_checkpoint')
 
             data : Data instance
                 raw data for analysis
