+++ idtxl/idtxl_io.py
@@ -3,10 +3,11 @@
 Provide functions to load and save IDTxl data, provide import functions (e.g.,
 mat-files, FieldTrip) and export functions (e.g., networkx, BrainNet Viewer).
 """
-# import json
+import json
 import pickle
 import h5py
 import networkx as nx
+from pprint import pprint
 import numpy as np
 import copy as cp
 import itertools as it
@@ -22,72 +23,58 @@ except ImportError as err:
          'https://pypi.python.org/pypi/networkx/2.0 to export and plot IDTxl '
          'results in this format.'))
 
-VERBOSE = False
+DEBUG = False
 
 
-# def save(data, file_path):
-#     """Save IDTxl data to disk.
+def save_json(d, file_path):
+    """Save dictionary to disk as JSON file.
 
-#     Save different data types to disk. Supported types are:
+    Writes dictionary to disk at the specified file path.
 
-#     - dictionaries with results, e.g., from MultivariateTE
-#     - numpy array
-#     - instance of IDTXL Data object
+    Args:
+        d : dict
+            dictionary to be written to disk
+        file_path : str
+            path to file (including extension)
 
-#     Note that while numpy arrays and Data instances are saved in binary for
-#     performance, dictionaries are saved in the json format, which is human-
-#     readable and also easily read into other programs (e.g., MATLAB:
-#     http://undocumentedmatlab.com/blog/json-matlab-integration).
+    Note: JSON does not recognize numpy data types, those are converted to
+    basic Python data types first.
+    """
+    data_json = _remove_numpy(d)
+    with open(file_path, 'w') as outfile:
+        json.dump(obj=data_json, fp=outfile, sort_keys=True)
 
-#     File extensions are
 
-#     - .txt for dictionaries (JSON file)
-#     - .npy for numpy array
-#     - .npz for Data instances
+def load_json(file_path):
+    """Load dictionary saved as JSON file from disk.
 
-#     If the extension is not provided in the file_path, the function will add
-#     it depending on the type of the data to be written.
+    Args:
+        file_path : str
+            path to file (including extension)
+    Returns:
+        dict
 
-#     Args:
-#         data : dict | numpy array | Data object
-#             data to be saved to disk
-#         file_path : string
-#             string with file name (including the path)
-#     """
-#     # Check if a file extension is provided in the file_path. Note that the
-#     # numpy save functions don't need an extension, they are added if
-#     # missing.
-#     if file_path.find('.', -4) == -1:
-#         add_extension = True
-#     else:
-#         add_extension = False
-
-#     if type(data) is dict:
-#         if add_extension:
-#             file_path = ''.join([file_path, '.txt'])
-#         # JSON does not recognize numpy arrays and data types, they have to
-#         # be converted before dumping them.
-#         data_json = _remove_numpy(data)
-#         if VERBOSE:
-#             print('writing file {0}'.format(file_path))
-#         with open(file_path, 'w') as outfile:
-#             json.dump(obj=data_json, fp=outfile, sort_keys=True)
-#     elif type(data) is np.ndarray:
-#         # TODO this can't handle scalars, handle this as an exception
-#         np.save(file_path, data)
-#     elif type(data) is __name__.data.Data:
-#         np.savez(file_path, data=data.data, normalised=data.normalise)
+    Note: JSON does not recognize numpy data structures and types. Numpy arrays
+    and data types (float, int) are thus converted to Python types and lists.
[…]
