### 2019-01-31 dc2dd9a
Add first version of checkpointing
Add first version of checkpointing. Add utilities to write JSON files to
save settings. Add utilities to write checkpoints, as well as data and
settings used. Add method to resume analysis from checkpoint. Add unit
tests and demo script.

Add functionality to manually add conditionals from a dictionary where
keys indicate targets, such that different conditionals can be added
for every target. This is used to resume analyses where already more than
a single target had been analyzed.

Fix minor error in adding of conditioning variables: check if a single
varia
[…]
