+++ idtxl/network_analysis.py
@@ -1,10 +1,16 @@
 """Parent class for network inference and network comparison.
 """
+import os.path
+from datetime import datetime
+from shutil import copyfile
+from pprint import pprint
+import ast
 import copy as cp
 import itertools as it
 import numpy as np
 from .estimator import find_estimator
 from . import idtxl_utils as utils
+from . import idtxl_io as io
 
 
 class NetworkAnalysis():
@@ -303,7 +309,12 @@ class NetworkAnalysis():
             cond = self.settings['add_conditionals']
             if type(cond) is tuple:  # easily add single variable
                 cond = [cond]
-            cond_idx = self._lag_to_idx(cond)            
+            elif type(cond) is dict:  # add conditioning variables per target
+                try:
+                    cond = cond[self.target]
+                except KeyError:
+                    return  # no additional variables for the current target
+            cond_idx = self._lag_to_idx(cond)
             candidate_set = list(set(candidate_set).difference(set(cond_idx)))
         return candidate_set
 
@@ -412,7 +423,7 @@ class NetworkAnalysis():
         Returns:
             numpy array
                 estimate of dependency measure for each link
-                
+
         Raises:
             ex.AlgorithmExhaustedError
                 Raised from estimate() when calculation cannot be made
@@ -500,3 +511,143 @@ class NetworkAnalysis():
                     conditional_realisations)
 
         return links
+
+    def _set_checkpointing_defaults(self, settings, data, sources, target):
+        """Set defaults for writing analysis checkpoints."""
+        settings.setdefault('write_ckp', False)
+        if settings['write_ckp']:
+            settings.setdefault('filename_ckp', './idtxl_checkpoint')
+            filename_ckp = '{0}.ckp'.format(settings['filename_ckp'])
+            if not os.path.isfile(filename_ckp):
+                self._initialise_checkpoint(settings, data, sources, target)
+            return settings
+        else:
+            return settings
+
+    def _initialise_checkpoint(self, settings, data, sources, targets):
+        """Write first checkpoint file, data, and settings to disk.
+
+        Called once at the beggining of an analysis using checkpointing. Write
+        data and analysis settings to disk. This needs to be done only once.
+        Initialise checkpoint file: write header with time stamp, path to data
+        and settings, and targets and sources to be analysed. The checkpoint
+        file is updated during the analyis.
+        """
+        # Check if targets is an int, convert to array.
+        if type(targets) is int:
+            targets = [targets]
+        # Write data to disk.
+        io.save_pickle(data,
+                       '{0}.dat'.format(settings['filename_ckp']))
+        # Write settings to disk.
+        io.save_json(settings,
+                     '{0}.json'.format(settings['filename_ckp']))
+
+        # Initialise checkpoint file for later updates.
+        filename_ckp = '{0}.ckp'.format(settings['filename_ckp'])
+        with open(filename_ckp, 'w') as text_file:
+            text_file.write('IDTxl checkpoint file.\n')
+            timestamp = datetime.now()
+            text_file.write('{:%Y-%m-%d %H:%M:%S}\n'.format(timestamp))
+            text_file.write('Raw data path: {}.dat\n'.format(
+                os.path.abspath(settings['filename_ckp'])))
+            text_file.write('Settings path: {}.json\n'.format(
+                os.path.abspath(settings['filename_ckp'])))
+            text_file.write('Targets to be analyzed: {}\n'.format(targets))
+            text_file.write('Sources to be analyzed: {}\n\n'.format(sources))
+            text_file.write(
+                'Selected variables (target: [sources]: [selected variables]):'
+                '\n{}'.format(targets[0]))
+
+    def _write_checkpoint(self):
+        """Write checkpoint to disk.
+
[…]
