### 2019-02-13 65711a4
Fix Kraskov reference in text
