### 2019-02-13 #1081 @pwollstadt
Thanks a lot! We found a minor error in the final proof, which I fixed in the repo's master (https://github.com/pwollstadt/IDTxl/commit/84e8469545402740ec5f89dc1b9dd925f955210). Is there a way to generate the final PDF from that?
