### 2019-02-13 #1081 @arfon
@pwollstadt - the `paper.md` metadata is currently broken. Please make this fix: https://github.com/pwollstadt/IDTxl/pull/31
