### 2019-02-14 #1081 @pwollstadt
Sorry about that. Thanks, I merged the fix into master (
https://github.com/pwollstadt/IDTxl/commit/a949f07857379727ddfab98c35074b0e8f580171).
