### 2019-02-14 #1081 @jlizier
Ah ok, I was trying to regenerate the proof / dry run after the merge, so we can verify it then give the deposit=true line. Could you do that pls @arfon ?
