### 2019-02-14 #1081 @arfon
@pwollstadt - please remove `http://arxiv.org/abs/1004.2515` from the DOI field in your bibtex. Something like this should work as an entry instead.

```
@misc{1004.2515,
Author = {Paul L. Williams and Randall D. Beer},
Title = {Nonnegative Decomposition of Multivariate Information},
Year = {2010},
Eprint = {arXiv:1004.2515},
}
```

Also, ignore the DOI suggestion here. Whedon does its best to suggest possible missing DOIs but sometimes it's wrong :-)
