### 2019-02-15 de4ed33
Fix arxiv reference
