### 2019-02-15 #1081 @pwollstadt
Thanks! :) I fixed the arxiv reference and left everything else as is in
https://github.com/pwollstadt/IDTxl/commit/de4ed33770311ba566c74d2f88fcafbecefcd6f5
