### 2019-02-17 #1081 @jlizier
The suggested missing DOI is for a preprint without a DOI, so I think we can proceed
